refactor: Move Firebase config files and update imports
Firebase configuration files have been moved to a different directory for better organization structure. Import paths in affected files have been updated accordingly to maintain functionality. Minor variable renaming for consistency in 'fire-config.js' was also performed in the process.
